appease the compiler by announcing this truncation is intended

+++ src/MinHook/trampoline.c
@@ -282,7 +282,7 @@ BOOL CreateTrampolineFunction(PTRAMPOLINE ct)
 #else
         __movsb((LPBYTE)ct->pTrampoline + newPos, (LPBYTE)pCopySrc, copySize);
 #endif
-        newPos += copySize;
+        newPos += (UINT8) copySize;
         oldPos += hs.len;
     }
     while (!finished);
